requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
cryptography
tqdm
numpy
torch
transformers